separate component from mdx
